requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==25.0
PyJWT==2.9.0
redis==6.1.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing-extensions==4.13.2
